i need you to create a tracking parcel page, once the user placed the order if there is a book in the cart.
means, after the payment is being done, we need to redirect user to a page like order-received. 

then in that page, we should show all the order details, and also need to add a tracking order
the value of this tracking order has to be a hyperlinked value , so we can redirect user to the tracking parcel page.