http://localhost:3000/learning-paths/beginner-property-investor http://localhost:3000/learning-paths/hdb-upgrader-strategist http://localhost:3000/learning-paths/condo-investment-specialist
need to redesign the learning paths. the path details container should add the total amount of the course fee. if Free means $0.00 for that course. 

Add something like a bundle information into the page as well. Telling them that the path is working as like a bundle deal with xx discounted offer.