i need you to design a page, to view the students profile from the public.
courses/making-the-right-move

there's a review from Gilbert Tan. User might doubt that whether this student is a real student or fake student (like us creating and making up a fake review). So it would be great if we allow user to see the student profile and see what courses did this person enrolled into, to verify if its a real or fake review. If this person never enroll the course but leave a good review or bad review, then user know this review may not be reliable. 
https://assembly.sg/profile/gilberttanjh/?view=student